Feat: useTouchArea hook 생성 및 적용
